inference py test

+++ inference.py
@@ -107,7 +107,7 @@ with torch.no_grad():
             X).reshape((int)(nheads / 4), -1, head_dim).repeat_interleave(4, dim=0)
 
         Attention = torch.matmul(
-            F.softmax(torch.matmul(Q, K.transpose(-1, -2)) / (head_dim ** 0.5), 1),
+            F.softmax(torch.matmul(Q, K.transpose(-1, -2)) / (head_dim ** 0.5), 1).half(),
             V
         )
